Move to views.py
