+++ parflow_data_management/scheduler/views.py
@@ -3,7 +3,7 @@ from django.shortcuts import render
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from ..models.cluster import Cluster
+from .models.cluster import Cluster
 from parflow_data_management.scheduler.tasks import remote_execute_cmd
 
 
@@ -15,6 +15,6 @@ def cluster_execute_page(request):
 def start_execution(request, cluster_id):
     try:
         remote_execute_cmd.delay(cluster_id, request.user.id, "ls")
-    except KeyError as e:
-        raise  # TODO: This should return an error code instead of propagate the exception
+    except KeyError:
+        raise
     return Response()
